+++ src/cache/inmemory/__tests__/readFromStore.ts
@@ -723,18 +723,21 @@ describe('reading from the store', () => {
 
     expect(missing).toEqual([
       new MissingFieldError(
-        `Can't find field 'missing' on object {
-  "present": "here"
-}`,
-        ["normal", "missing"],
-        query,
-        {}, // variables
-      ),
-      new MissingFieldError(
-        `Can't find field 'missing' on object {
-  "present": "also here"
-}`,
-        ["clientOnly", "missing"],
+        `Can't find field 'missing' on object ${JSON.stringify({
+          present: "also here",
+        }, null, 2)}`,
+        {
+          normal: {
+            missing: `Can't find field 'missing' on object ${
+              JSON.stringify({ present: "here" }, null, 2)
+            }`,
+          },
+          clientOnly: {
+            missing: `Can't find field 'missing' on object ${
+              JSON.stringify({ present: "also here" }, null, 2)
+            }`,
+          },
+        },
         query,
         {}, // variables
       ),
@@ -1289,8 +1292,19 @@ describe('reading from the store', () => {
     expect(diffChickens()).toEqual({
       complete: false,
       missing: [
-        expect.anything(),
-        expect.anything(),
+        new MissingFieldError(
+          "Can't find field 'id' on object {}",
+          {
+            chickens: {
+              1: {
+                id: "Can't find field 'id' on object {}",
+                inCoop: "Can't find field 'inCoop' on object {}",
+              },
+            },
+          },
+          expect.anything(), // query
+          expect.anything(), // variables
+        ),
       ],
       result: {
         chickens: [